requirements.txt
streamlit=1.47.0
requests=2.32.4
Pillow=11.3.0
python-dotenv=1.1.1